Move page/sector size constants from gimlet-hf-api to new qspi-api crate
