+++ drv/gimlet-hf-api/Cargo.toml
@@ -10,6 +10,7 @@ derive-idol-err = {path = "../../lib/derive-idol-err" }
 userlib = {path = "../../sys/userlib"}
 num-traits = { version = "0.2.12", default-features = false }
 zerocopy = "0.6.1"
+drv-qspi-api = {path = "../qspi-api"}
 drv-hash-api = {path = "../hash-api"}
 
 [build-dependencies]
