+++ drv/gimlet-hf-api/src/lib.rs
@@ -11,6 +11,8 @@ use drv_hash_api::SHA256_SZ;
 use userlib::*;
 use zerocopy::AsBytes;
 
+pub use drv_qspi_api::{PAGE_SIZE_BYTES, SECTOR_SIZE_BYTES};
+
 /// Errors that can be produced from the host flash server API.
 ///
 /// This enumeration doesn't include errors that result from configuration
@@ -44,17 +46,5 @@ pub enum HfDevSelect {
     Flash0 = 0,
     Flash1 = 1,
 }
-
-/// Size in bytes of a single page of data (i.e., the max length of slice we
-/// accept for `page_program()` and `read()`).
-// Note: There is no static check that this matches what's in our idl file in
-// terms of _client_ generation, but the server can use this constant in its
-// trait impl, which will produce a compile-time error if it doesn't match the
-// length in the idl file.
-pub const PAGE_SIZE_BYTES: usize = 256;
-
-/// Size in bytes of a single sector of data (i.e., the size of the data erased
-/// by a call to `sector_erase()`).
-pub const SECTOR_SIZE_BYTES: usize = 65_536;
 
 include!(concat!(env!("OUT_DIR"), "/client_stub.rs"));
