+++ drv/qspi-api/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "drv-qspi-api"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+
+# This section is here to discourage RLS/rust-analyzer from doing test builds,
+# since test builds don't work for cross compilation.
+[lib]
+test = false
+bench = false
