+++ drv/stm32h7-qspi/Cargo.toml
@@ -3,9 +3,8 @@ name = "drv-stm32h7-qspi"
 version = "0.1.0"
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
+drv-qspi-api = {path = "../qspi-api"}
 stm32h7 = { version = "0.14", default-features = false }
 vcell = "0.1.2"
 zerocopy = "0.6.1"
